feat: create asyncPersistance
