+++ package.json
@@ -9,8 +9,10 @@
   "scripts": {
     "lint": "eslint",
     "lint:fix": "eslint --fix",
-    "unit": "node --test test.js",
+    "unit": "node --test test/test.js",
     "test": "npm run lint && npm run unit",
+    "coverage": "c8 --reporter=lcov npm run unit",
+    "coverage:report": "c8 report",
     "release": "read -p 'GITHUB_TOKEN: ' GITHUB_TOKEN && export GITHUB_TOKEN=$GITHUB_TOKEN && release-it --disable-metrics"
   },
   "release-it": {
@@ -50,11 +52,12 @@
   "devDependencies": {
     "@fastify/pre-commit": "^2.2.0",
     "aedes": "^0.51.3",
-    "aedes-persistence": "^10.0.0",
-    "eslint": "^9.21.0",
-    "level": "^9.0.0",
+    "aedes-persistence": "^10.1.0",
+    "c8": "^10.1.3",
+    "eslint": "^9.27.0",
+    "level": "^10.0.0",
     "neostandard": "^0.12.1",
-    "release-it": "^18.1.2"
+    "release-it": "^19.0.2"
   },
   "dependencies": {
     "aedes-packet": "^3.0.0",
